requirements/base.txt
django == 3.1
pillow == 8.1.0
psycopg2 == 2.8.6
djangorestframework == 3.11.0
django-cors-headers == 3.7.0
twilio ==5.7.0
celery == 5.0.5
django-environ==0.4.5
django-celery-results ==2.0.1
django-celery-beat == 2.2.0
django-filter==2.4.0